requirements.txt
torch
torchvision
torchaudio
transformers
sentence-transformers
faiss-cpu
datasets
pyyaml
tqdm
numpy
